Installation independent of previous install
Before this update, you needed to include a ".install" file in a
specific directory to invoke the installer. Now the application will
look for a specific file generated by an installer
(protected/config/main-ext.php) which has the database connection
details. If this file does not exist (cannot connect to DB) then it will
invoke the installer automatically.

To reinstall the application, you will need to delete the main-ext.php
file from the configuration directory.