Watchlist and Cron implementation
Users can now add keywords to their watchlist. Every time a new post is
posted a Cron job will go through everyone's watchlist and email them if
the post description has a keyword in their watchlist. Cron will run in
the background after a post is submitted so as not to bog down the
user's experience.