Overhaul on updating table schemas
This major update will make sure the system is healthy. This includes
checking the Yii framework, checking the database exists, and verifying
the table schemas match for each appropriate object. If they do not
match the system will automatically update the database so that they do
match.

Basically all database changes are done through their associated objects
and their schemas. The reason for this is so that programmers do not
have to manually change their tables whenever an update is pushed.