Smoothed out installation procedures
Installation instructions are now unnecessary.
The installation first creates the extra config file with the database
connections. It then redirects to the main page where the
"beforeAction()" event is executed before the page is rendered. This
action checks the system for any database/table changes, and makes the
corrections automatically. In this way, it will actually automatically
install the tables.

All a user needs to do now is download this as a zip, extract, and load
the application in a web-page. The rest of the process should be
automatic. :)